move tests to cuda_test.cc
