+++ cpp/src/arrow/c/CMakeLists.txt
@@ -15,23 +15,7 @@
 # specific language governing permissions and limitations
 # under the License.
 
-if(ARROW_CUDA)
-  if(ARROW_BUILD_SHARED)
-    set(ARROW_CUDA_LIBRARY arrow_cuda_shared)
-  else()
-    set(ARROW_CUDA_LIBRARY arrow_cuda_static)
-  endif()
-
-  set(ARROW_C_TEST_LINK_LIBS ${ARROW_CUDA_LIBRARY} ${ARROW_TEST_LINK_LIBS})
-else()
-  set(ARROW_C_TEST_LINK_LIBS ${ARROW_TEST_LINK_LIBS})
-endif()
-
-add_arrow_test(bridge_test
-               PREFIX
-               "arrow-c"
-               STATIC_LINK_LIBS
-               ${ARROW_C_TEST_LINK_LIBS})
+add_arrow_test(bridge_test PREFIX "arrow-c")
 add_arrow_test(dlpack_test)
 
 add_arrow_benchmark(bridge_benchmark)
